+++ hydra-cluster/src/Hydra/Cluster/Scenarios.hs
@@ -118,6 +118,7 @@ import Hydra.Ledger.Cardano (mkSimpleTx, mkTransferTx, unsafeBuildTransaction)
 import Hydra.Ledger.Cardano.Evaluate (maxTxExecutionUnits)
 import Hydra.Logging (Tracer, traceWith)
 import Hydra.Node.DepositPeriod (DepositPeriod (..))
+import Hydra.Node.State (SyncedStatus (..))
 import Hydra.Options (CardanoChainConfig (..), ChainBackendOptions (..), DirectOptions (..), RunOptions (..), startChainFrom)
 import Hydra.Tx (HeadId (..), IsTx (balance), Party, headIdToCurrencySymbol, txId)
 import Hydra.Tx.ContestationPeriod qualified as CP
@@ -144,7 +145,8 @@ import HydraNode (
   waitNoMatch,
   withHydraCluster,
   withHydraNode,
-  withPreparedHydraNode,
+  withHydraNodeCatchingUp,
+  withPreparedHydraNodeInSync,
  )
 import Network.HTTP.Conduit (parseUrlThrow)
 import Network.HTTP.Conduit qualified as L
@@ -539,7 +541,7 @@ singlePartyOpenAHead tracer workDir backend hydraScriptsTxId persistenceRotateAf
     let hydraTracer = contramap FromHydraNode tracer
     options <- prepareHydraNode aliceChainConfig workDir 1 aliceSk [] [] id
     let options' = options{persistenceRotateAfter}
-    withPreparedHydraNode hydraTracer workDir 1 options' $ \n1 -> do
+    withPreparedHydraNodeInSync hydraTracer workDir 1 options' $ \n1 -> do
       -- Initialize & open head
       send n1 $ input "Init" []
       blockTime <- Backend.getBlockTime backend
@@ -1403,7 +1405,6 @@ initWithWrongKeys workDir tracer backend hydraScriptsTxId = do
   withHydraNode hydraTracer aliceChainConfig workDir 3 aliceSk [bobVk] [3, 4] $ \n1 -> do
     withHydraNode hydraTracer bobChainConfig workDir 4 bobSk [aliceVk] [3, 4] $ \n2 -> do
       seedFromFaucet_ backend aliceCardanoVk 100_000_000 (contramap FromFaucet tracer)
-
       send n1 $ input "Init" []
       headId <-
         waitForAllMatch 10 [n1] $
@@ -1668,7 +1669,7 @@ rejectCommit tracer workDir blockTime backend hydraScriptsTxId =
     let pparamsDecorator = atKey "utxoCostPerByte" ?~ toJSON (Aeson.Number 4310)
     optionsWithUTxOCostPerByte <- prepareHydraNode aliceChainConfig workDir 1 aliceSk [] [] pparamsDecorator
 
-    withPreparedHydraNode hydraTracer workDir 1 optionsWithUTxOCostPerByte $ \n1 -> do
+    withPreparedHydraNodeInSync hydraTracer workDir 1 optionsWithUTxOCostPerByte $ \n1 -> do
       send n1 $ input "Init" []
       headId <- waitMatch (10 * blockTime) n1 $ headIsInitializingWith (Set.fromList [alice])
 
@@ -2107,7 +2108,7 @@ canSideLoadSnapshot tracer workDir backend hydraScriptsTxId = do
   seedFromFaucet_ backend bobCardanoVk 100_000_000 (contramap FromFaucet tracer)
   seedFromFaucet_ backend carolCardanoVk 100_000_000 (contramap FromFaucet tracer)
   blockTime <- Backend.getBlockTime backend
-  let contestationPeriod = 1
+  let contestationPeriod = 1000
 
   networkId <- Backend.queryNetworkId backend
   aliceChainConfig <-
@@ -2126,7 +2127,7 @@ canSideLoadSnapshot tracer workDir backend hydraScriptsTxId = do
       -- Carol starts its node misconfigured
       let pparamsDecorator = atKey "maxTxSize" ?~ toJSON (Aeson.Number 0)
       wrongOptions <- prepareHydraNode carolChainConfig workDir 3 carolSk [aliceVk, bobVk] [1, 2, 3] pparamsDecorator
-      tx <- withPreparedHydraNode hydraTracer workDir 3 wrongOptions $ \n3 -> do
+      tx <- withPreparedHydraNodeInSync hydraTracer workDir 3 wrongOptions $ \n3 -> do
         send n1 $ input "Init" []
         headId <- waitForAllMatch (10 * blockTime) [n1, n2, n3] $ headIsInitializingWith (Set.fromList [alice, bob, carol])
 
@@ -2146,12 +2147,12 @@ canSideLoadSnapshot tracer workDir backend hydraScriptsTxId = do
         send n1 $ input "NewTx" ["transaction" .= tx]
 
         -- Alice and Bob accept it
-        waitForAllMatch (200 * blockTime) [n1, n2] $ \v -> do
+        waitForAllMatch (20 * blockTime) [n1, n2] $ \v -> do
           guard $ v ^? key "tag" == Just "TxValid"
           guard $ v ^? key "transactionId" == Just (toJSON $ txId tx)
 
         -- Carol does not because of its node being misconfigured
-        waitMatch (10 * blockTime) n3 $ \v -> do
+        waitMatch (20 * blockTime) n3 $ \v -> do
           guard $ v ^? key "tag" == Just "TxInvalid"
           guard $ v ^? key "transaction" . key "txId" == Just (toJSON $ txId tx)
 
@@ -2165,7 +2166,7 @@ canSideLoadSnapshot tracer workDir backend hydraScriptsTxId = do
         pure tx
 
       -- Carol disconnects and the others observe it
-      waitForAllMatch (100 * blockTime) [n1, n2] $ \v -> do
+      waitForAllMatch 5 [n1, n2] $ \v -> do
         guard $ v ^? key "tag" == Just "PeerDisconnected"
 
       -- Carol reconnects with healthy reconfigured node
@@ -2173,11 +2174,11 @@ canSideLoadSnapshot tracer workDir backend hydraScriptsTxId = do
         -- Carol re-submits the same transaction
         send n3 $ input "NewTx" ["transaction" .= tx]
         -- Carol accepts it
-        waitMatch (10 * blockTime) n3 $ \v -> do
+        waitMatch (20 * blockTime) n3 $ \v -> do
           guard $ v ^? key "tag" == Just "TxValid"
           guard $ v ^? key "transactionId" == Just (toJSON $ txId tx)
         -- But now Alice and Bob does not because they already applied it
-        waitForAllMatch (200 * blockTime) [n1, n2] $ \v -> do
+        waitForAllMatch (20 * blockTime) [n1, n2] $ \v -> do
           guard $ v ^? key "tag" == Just "TxInvalid"
           guard $ v ^? key "transaction" . key "txId" == Just (toJSON $ txId tx)
 
@@ -2247,12 +2248,12 @@ canResumeOnMemberAlreadyBootstrapped tracer workDir backend hydraScriptsTxId = d
       waitMatch (20 * blockTime) n2 $ \v -> do
         guard $ v ^? key "tag" == Just "Greetings"
         guard $ v ^? key "headStatus" == Just (toJSON Idle)
-
       threadDelay 5
 
     callProcess "rm" ["-rf", workDir </> "state-2"]
+    threadDelay 1
 
-    withHydraNode hydraTracer bobChainConfig workDir 2 bobSk [aliceVk] [1, 2] (const $ pure ())
+    withHydraNode hydraTracer bobChainConfig workDir 2 bobSk [aliceVk] [1, 2] (const $ threadDelay 1)
       `shouldThrow` \(e :: SomeException) ->
         "hydra-node" `isInfixOf` show e
           && "etcd" `isInfixOf` show e
@@ -2260,6 +2261,94 @@ canResumeOnMemberAlreadyBootstrapped tracer workDir backend hydraScriptsTxId = d
     setEnv "ETCD_INITIAL_CLUSTER_STATE" "existing"
     withHydraNode hydraTracer bobChainConfig workDir 2 bobSk [aliceVk] [1, 2] (const $ pure ())
     unsetEnv "ETCD_INITIAL_CLUSTER_STATE"
+ where
+  hydraTracer = contramap FromHydraNode tracer
+
+-- XXX: restart scenarios require 3 party cluster in order to observe PeerDisconnected instead of NetworkDisconnected
+waitsForChainInSyncAndSecure :: ChainBackend backend => Tracer IO EndToEndLog -> FilePath -> backend -> [TxId] -> IO ()
+waitsForChainInSyncAndSecure tracer workDir backend hydraScriptsTxId = do
+  let clients = [Alice, Bob, Carol]
+  [(aliceCardanoVk, _), (bobCardanoVk, _), (carolCardanoVk, carolCardanoSk)] <- forM clients keysFor
+  seedFromFaucet_ backend aliceCardanoVk 100_000_000 (contramap FromFaucet tracer)
+  seedFromFaucet_ backend bobCardanoVk 100_000_000 (contramap FromFaucet tracer)
+  seedFromFaucet_ backend carolCardanoVk 100_000_000 (contramap FromFaucet tracer)
+
+  blockTime <- Backend.getBlockTime backend
+  networkId <- Backend.queryNetworkId backend
+  contestationPeriod <- CP.fromNominalDiffTime (100 * blockTime)
+  aliceChainConfig <-
+    chainConfigFor Alice workDir backend hydraScriptsTxId [Bob, Carol] contestationPeriod
+      <&> setNetworkId networkId
+  bobChainConfig <-
+    chainConfigFor Bob workDir backend hydraScriptsTxId [Alice, Carol] contestationPeriod
+      <&> setNetworkId networkId
+  carolChainConfig <-
+    chainConfigFor Carol workDir backend hydraScriptsTxId [Alice, Bob] contestationPeriod
+      <&> setNetworkId networkId
+
+  withHydraNode hydraTracer aliceChainConfig workDir 1 aliceSk [bobVk, carolVk] [1, 2, 3] $ \n1 -> do
+    withHydraNode hydraTracer bobChainConfig workDir 2 bobSk [aliceVk, carolVk] [1, 2, 3] $ \n2 -> do
+      -- Open a head while Carol online
+      headId <- withHydraNode hydraTracer carolChainConfig workDir 3 carolSk [aliceVk, bobVk] [1, 2, 3] $ \n3 -> do
+        send n1 $ input "Init" []
+        headId <- waitForAllMatch (10 * blockTime) [n1, n2, n3] $ headIsInitializingWith (Set.fromList [alice, bob, carol])
+
+        -- Alice commits nothing
+        requestCommitTx n1 mempty >>= Backend.submitTransaction backend
+        -- Bob commits nothing
+        requestCommitTx n2 mempty >>= Backend.submitTransaction backend
+        -- Carol commits something
+        carolUTxO <- seedFromFaucet backend carolCardanoVk (lovelaceToValue 1_000_000) (contramap FromFaucet tracer)
+        requestCommitTx n3 carolUTxO >>= Backend.submitTransaction backend
+
+        -- Observe open with the relevant UTxOs
+        waitFor hydraTracer (20 * blockTime) [n1, n2, n3] $
+          output "HeadIsOpen" ["utxo" .= toJSON carolUTxO, "headId" .= headId]
+
+        pure headId
+
+      -- Carol disconnects and the others observe it
+      waitForAllMatch 5 [n1, n2] $ \v -> do
+        guard $ v ^? key "tag" == Just "PeerDisconnected"
+
+      -- Wait for some blocks to roll forward
+      threadDelay $ realToFrac (CP.unsyncedPolicy contestationPeriod + 50 * blockTime)
+
+      -- Alice closes the head while Carol offline
+      send n1 $ input "Close" []
+      waitForAllMatch (20 * blockTime) [n1, n2] $ \v -> do
+        guard $ v ^? key "tag" == Just "HeadIsClosed"
+        guard $ v ^? key "headId" == Just (toJSON headId)
+
+      -- Carol restarts
+      withHydraNodeCatchingUp hydraTracer carolChainConfig workDir 3 carolSk [aliceVk, bobVk] [1, 2, 3] $ \n3 -> do
+        -- The node reports that it is in the Open state and out of sync with the chain
+        waitMatch 5 n3 $ \v -> do
+          guard $ v ^? key "tag" == Just "Greetings"
+          guard $ v ^? key "headStatus" == Just (toJSON Open)
+          guard $ v ^? key "me" == Just (toJSON carol)
+          guard $ v ^? key "chainSyncedStatus" == Just (toJSON CatchingUp)
+          guard $ isJust (v ^? key "hydraNodeVersion")
+
+        -- Then, Carol attempts submits a new transaction,
+        -- without waiting for head to be closed
+        utxo <- getSnapshotUTxO n3
+        tx <- mkTransferTx testNetworkId utxo carolCardanoSk carolCardanoVk
+        send n3 $ input "NewTx" ["transaction" .= tx]
+
+        waitMatch 5 n3 $ \v -> do
+          guard $ v ^? key "tag" == Just "RejectedInput"
+          guard $ v ^? key "reason" == Just "chain out of sync"
+
+        -- Carol API notifies the node is back on sync with the chain
+        -- note this is tuned based on how long it takes to sync
+        waitMatch (CP.unsyncedPolicy contestationPeriod + 5 * blockTime) n3 $ \v -> do
+          guard $ v ^? key "tag" == Just "NodeSynced"
+
+        -- Finally, Carol observes the head getting closed
+        waitMatch (20 * blockTime) n3 $ \v -> do
+          guard $ v ^? key "tag" == Just "HeadIsClosed"
+          guard $ v ^? key "headId" == Just (toJSON headId)
  where
   hydraTracer = contramap FromHydraNode tracer