+++ hydra-node/src/Hydra/API/Server.hs
@@ -116,6 +116,7 @@ withAPIServer config env stateFile party eventSource tracer chain pparams server
                   update nodeStateP stateChanged
                   update commitInfoP stateChanged
                   update pendingDepositsP stateChanged
+                  update networkInfoP stateChanged
             )
     (notifyServerRunning, waitForServerRunning) <- setupServerNotification
 
@@ -265,11 +266,13 @@ mkTimedServerOutputFromStateEvent event =
     StateChanged.TickObserved{} -> Nothing
     StateChanged.LocalStateCleared{..} -> Just SnapshotSideLoaded{..}
     StateChanged.Checkpoint{state} -> Just $ EventLogRotated state
+    StateChanged.NodeUnsynced -> Just NodeUnsynced
+    StateChanged.NodeSynced -> Just NodeSynced
 
 -- | Projection to obtain the list of pending deposits.
 projectPendingDeposits :: IsTx tx => [TxIdType tx] -> StateChanged.StateChanged tx -> [TxIdType tx]
 projectPendingDeposits txIds = \case
-  StateChanged.Checkpoint{state = NodeState{pendingDeposits}} -> Map.keys pendingDeposits
+  StateChanged.Checkpoint{state} -> Map.keys (pendingDeposits state)
   StateChanged.DepositRecorded{depositTxId} -> depositTxId : txIds
   StateChanged.DepositRecovered{depositTxId} -> filter (/= depositTxId) txIds
   StateChanged.CommitFinalized{depositTxId} -> filter (/= depositTxId) txIds
@@ -280,10 +283,11 @@ projectPendingDeposits txIds = \case
 -- state since this is when Head parties need to commit some funds.
 projectCommitInfo :: CommitInfo -> StateChanged.StateChanged tx -> CommitInfo
 projectCommitInfo commitInfo = \case
-  StateChanged.Checkpoint NodeState{headState = state} -> case state of
-    Initial InitialState{headId} -> NormalCommit headId
-    Open OpenState{headId} -> IncrementalCommit headId
-    _ -> CannotCommit
+  StateChanged.Checkpoint state ->
+    case headState state of
+      Initial InitialState{headId} -> NormalCommit headId
+      Open OpenState{headId} -> IncrementalCommit headId
+      _ -> CannotCommit
   StateChanged.HeadInitialized{headId} -> NormalCommit headId
   StateChanged.HeadOpened{headId} -> IncrementalCommit headId
   StateChanged.HeadAborted{} -> CannotCommit