+++ hydra-node/src/Hydra/API/ServerOutput.hs
@@ -17,7 +17,7 @@ import Hydra.HeadLogic.State qualified as HeadState
 import Hydra.Ledger (ValidationError)
 import Hydra.Network (Host, ProtocolVersion)
 import Hydra.Node.Environment (Environment (..))
-import Hydra.Node.State (NodeState)
+import Hydra.Node.State (NodeState, SyncedStatus)
 import Hydra.Prelude hiding (seq)
 import Hydra.Tx (HeadId, Party, Snapshot, SnapshotNumber, getSnapshot)
 import Hydra.Tx qualified as Tx
@@ -74,6 +74,7 @@ instance ArbitraryIsTx tx => Arbitrary (DecommitInvalidReason tx) where
 data ClientMessage tx
   = CommandFailed {clientInput :: ClientInput tx, state :: HeadState tx}
   | PostTxOnChainFailed {postChainTx :: PostChainTx tx, postTxError :: PostTxError tx}
+  | RejectedInput {clientInput :: ClientInput tx, reason :: Text}
   deriving (Eq, Show, Generic)
 
 instance IsChainState tx => ToJSON (ClientMessage tx) where
@@ -105,6 +106,7 @@ data Greetings tx = Greetings
   , hydraNodeVersion :: String
   , env :: Environment
   , networkInfo :: NetworkInfo
+  , chainSyncedStatus :: SyncedStatus
   }
   deriving (Generic)
 
@@ -215,6 +217,8 @@ data ServerOutput tx
     -- Any signing round has been discarded, and the snapshot leader has changed accordingly.
     SnapshotSideLoaded {headId :: HeadId, snapshotNumber :: SnapshotNumber}
   | EventLogRotated {checkpoint :: NodeState tx}
+  | NodeUnsynced
+  | NodeSynced
   deriving stock (Generic)
 
 deriving stock instance IsChainState tx => Eq (ServerOutput tx)
@@ -288,6 +292,8 @@ prepareServerOutput config response =
     PeerDisconnected{} -> encodedResponse
     SnapshotSideLoaded{} -> encodedResponse
     EventLogRotated{} -> encodedResponse
+    NodeUnsynced{} -> encodedResponse
+    NodeSynced{} -> encodedResponse
  where
   encodedResponse = encode response